requirements.txt
clipboard==0.0.4
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
plyer==2.1.0
pyperclip==1.8.2
pytesseract==0.3.10
